refactor(relay): extract the NATS relay publisher to a shared package
Replace the e2e's inlined copy of the relay publisher (the one place the
migration duplicated real production LOGIC, not just a wire shape) with a single
source of truth in @decocms/sandbox/dispatch/relay-publisher — next to the
relayLineSchema it already depends on.

- New packages/sandbox/dispatch/relay-publisher.ts owns the wire format (subject
  + msgId scheme + checkpoint debounce + the publisher + publishRelayBodyToNats).
  It is metrics-agnostic; OpenTelemetry is injected via optional `hooks`.
- apps/mesh/src/link-daemon/direct-nats-publisher.ts becomes a thin adapter that
  injects the app's metrics via hooks and re-exports the rest — its importers
  (handle-local-dispatch, dispatch-link-work-item) are unchanged.
- projector-stream-messages.ts sources the build helpers from the package and
  re-exports them, so the projector/ingest consumers are untouched.
- packages/e2e/fixtures/relay-nats.ts imports the package; the inlined
  fixtures/relay-publisher.ts is deleted.
- The publisher's unit test moves to the package (its new home).

This also fixes drift the inline had already accrued: since it was copied, the
real publisher gained an `onPublished` callback + metrics (via the main merge) —
the e2e copy was stale. Extraction makes that class of drift impossible.

Verified: sandbox/mesh/e2e tsc 0, lint 0, knip 0; publisher (10) + projector
(15) + dispatch (52) unit tests pass.
