+++ packages/sandbox/dispatch/relay-publisher.ts
@@ -1,36 +1,39 @@
 /**
- * Direct-NATS relay publisher — a black-box copy of what a real desktop daemon
- * does (`src/link-daemon/direct-nats-publisher.ts`), inlined so the e2e suite
- * owns the wire format (JetStream subject + msgId scheme + checkpoint debounce)
- * instead of importing app source. A divergence from the app's projector-stream
- * format is a wire-contract regression signal the suite SHOULD catch.
+ * Direct-NATS relay publisher — the producer half of the chunk-relay protocol
+ * (spec: "Transport Convergence", protocol v2). A desktop daemon publishes each
+ * seq-numbered RelayLine straight to the JetStream subject
+ * `decopilot.stream.<runId>`; the always-on durable projector consumes it.
  *
- * The relay-line schema is the published contract `@decocms/sandbox/dispatch`,
- * so that one is imported, not duplicated.
+ * This is the single source of truth for the wire format (subject + msgId
+ * scheme + checkpoint debounce + envelope shapes), shared by:
+ *   - the app's live producer (apps/mesh/src/link-daemon/direct-nats-publisher.ts,
+ *     a thin adapter that injects OpenTelemetry metrics via `hooks`), and
+ *   - the e2e "fake daemon" relay (packages/e2e/fixtures/relay-nats.ts).
+ *
+ * It is metrics-agnostic on purpose — observability is injected via optional
+ * `hooks` so a test (no OTel meter wired) can drive it directly.
  */
 import type { JetStreamClient } from "@nats-io/jetstream";
-import {
-  relayLineSchema,
-  type RelayLine,
-} from "@decocms/sandbox/dispatch/relay";
+import { relayLineSchema, type RelayLine } from "./relay";
 
-// --- Stream subject + msgId scheme (mirrors projector-stream-messages.ts) -----
+// --- Stream subject + msgId scheme -------------------------------------------
 
-const DECOPILOT_STREAM_SUBJECT_PREFIX = "decopilot.stream";
+export const DECOPILOT_STREAM_SUBJECT_PREFIX = "decopilot.stream";
 
-/** Debounce between incremental checkpoint markers (matches the app producer). */
-const CHECKPOINT_DEBOUNCE_MS = 3000;
+/** Debounce between incremental checkpoint markers, shared by BOTH producers
+ *  (hosted ingest + desktop daemon relay). Single source of truth. */
+export const CHECKPOINT_DEBOUNCE_MS = 3000;
 
 function assertSafeSubjectToken(id: string): void {
   if (/[.*>\s]/.test(id)) throw new Error("Invalid NATS subject token");
 }
 
-function streamSubject(runId: string): string {
+export function streamSubject(runId: string): string {
   assertSafeSubjectToken(runId);
   return `${DECOPILOT_STREAM_SUBJECT_PREFIX}.${runId}`;
 }
 
-function buildChunkMsgId(input: {
+export function buildChunkMsgId(input: {
   runId: string;
   fenceToken: string;
   seq: number;
@@ -42,7 +45,7 @@ function buildChunkMsgId(input: {
     : `${base}:frag:${input.fragmentIndex}`;
 }
 
-function buildDoneMsgId(input: {
+export function buildDoneMsgId(input: {
   runId: string;
   fenceToken: string;
   finalSeq: number;
@@ -50,7 +53,7 @@ function buildDoneMsgId(input: {
   return `${input.runId}:${input.fenceToken}:done:${input.finalSeq}`;
 }
 
-function buildCheckpointMsgId(input: {
+export function buildCheckpointMsgId(input: {
   runId: string;
   fenceToken: string;
   headSeq: number;
@@ -59,6 +62,14 @@ function buildCheckpointMsgId(input: {
 }
 
 // --- Publisher ----------------------------------------------------------------
+
+/** Observability injected by the app's live producer; omitted by tests. */
+export interface PublisherHooks {
+  /** Encode duration (ms) for each published ui-message-chunk. */
+  onChunkEncoded?: (ms: number) => void;
+  /** Fired once per published ui-message-chunk (count delta = 1). */
+  onChunkPublished?: () => void;
+}
 
 export interface DirectNatsPublisher {
   publishLine(input: {
@@ -80,16 +91,20 @@ export interface DirectNatsPublisher {
 
 export function createDirectNatsPublisher(input: {
   js: Pick<JetStreamClient, "publish">;
+  hooks?: PublisherHooks;
 }): DirectNatsPublisher {
   const encoder = new TextEncoder();
+  const hooks = input.hooks;
   return {
     async publishLine({ runId, fenceToken, line }) {
       if (line.event.type === "ui-message-chunk") {
-        await input.js.publish(
-          streamSubject(runId),
-          encoder.encode(JSON.stringify({ p: line.event.chunk })),
-          { msgID: buildChunkMsgId({ runId, fenceToken, seq: line.seq }) },
-        );
+        const t0 = performance.now();
+        const bytes = encoder.encode(JSON.stringify({ p: line.event.chunk }));
+        hooks?.onChunkEncoded?.(performance.now() - t0);
+        hooks?.onChunkPublished?.();
+        await input.js.publish(streamSubject(runId), bytes, {
+          msgID: buildChunkMsgId({ runId, fenceToken, seq: line.seq }),
+        });
         return "published";
       }
       if (line.event.type === "error") {
@@ -121,7 +136,9 @@ export function createDirectNatsPublisher(input: {
 }
 
 export interface PublishRelayBodyResult {
-  /** Highest wireSeq observed across ALL lines (incl. the terminal `done`). */
+  /** Highest wireSeq observed across ALL lines (incl. the terminal `done`
+   *  line). The chunk-relay poster checks `lastSeq >= terminal seq`, so this
+   *  must be the done line's seq, not the content count. */
   lastSeq: number;
 }
 
@@ -149,6 +166,7 @@ async function* ndjsonValues(
       if (done) break;
       buffer += decoder.decode(value, { stream: true });
       const lines = buffer.split("\n");
+      // keep the last (possibly incomplete) chunk in buffer
       buffer = lines.pop() ?? "";
       for (const line of lines) {
         const trimmed = line.trim();
@@ -160,6 +178,7 @@ async function* ndjsonValues(
         }
       }
     }
+    // flush remaining
     const flushed = buffer + decoder.decode();
     const trimmed = flushed.trim();
     if (trimmed !== "") {
@@ -181,15 +200,30 @@ export async function publishRelayBodyToNats(input: {
   publisher: DirectNatsPublisher;
   now?: () => number;
   checkpointDebounceMs?: number;
+  /**
+   * Reports the highest CONTENT seq durably published so far (each `publishLine`
+   * awaits its JetStream PubAck before `finalSeq` advances, so everything up to
+   * this value is durable). The relay uses it to drop the confirmed prefix from
+   * the outbox. Throttled to the checkpoint cadence (+ a final report) so it
+   * does not trigger a truncation per line.
+   */
+  onPublished?: (seq: number) => void;
 }): Promise<PublishRelayBodyResult> {
   const now = input.now ?? Date.now;
   const debounceMs = input.checkpointDebounceMs ?? CHECKPOINT_DEBOUNCE_MS;
   let maxWireSeq = 0;
-  let finalSeq = 0;
-  let lastCheckpointAt = now();
+  let finalSeq = 0; // highest CONTENT seq (chunks + converted error), for done
+  let reportedSeq = 0;
+  let lastCheckpointAt = now(); // first checkpoint fires one debounce window in
+  const reportPublished = (): void => {
+    if (input.onPublished && finalSeq > reportedSeq) {
+      reportedSeq = finalSeq;
+      input.onPublished(finalSeq);
+    }
+  };
   for await (const value of ndjsonValues(input.body)) {
     const parsed = relayLineSchema.safeParse(value);
-    if (!parsed.success) continue;
+    if (!parsed.success) continue; // skip blank/garbage lines (heartbeats already skipped)
     const line = parsed.data;
     maxWireSeq = Math.max(maxWireSeq, line.seq);
     const result = await input.publisher.publishLine({
@@ -206,6 +240,7 @@ export async function publishRelayBodyToNats(input: {
           fenceToken: input.fenceToken,
           headSeq: finalSeq,
         });
+        reportPublished();
       }
     }
   }
@@ -216,5 +251,6 @@ export async function publishRelayBodyToNats(input: {
       finalSeq,
     });
   }
+  reportPublished();
   return { lastSeq: maxWireSeq };
 }
